requirements.txt
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
et_xmlfile==2.0.0
Flask==3.1.2
fonttools==4.59.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
